Add Keystone admin/public to enabled services list
These services need to be added to the ServiceNames resources list of
enabled services per role if we want groupings for them to be created in
the ansible inventory.

This same logic already exists in network/ports/net_ip_list_map.j2.yaml,
where the hieradata for <service>_node_[names,ips] are created. We also
need these service groupings to exist in the ansible inventory if we
want to make this hieradata generic instead of hardcoded coming out of
the heat templates.

Change-Id: Ie4c221a850a9018b4eb62bfd3d54ac22395b7a9c