Remove ceph-ansible fetch directory after it is saved
Because the ceph-ansible fetch directory is backed up to Swift
or the directory defined by LocalCephAnsibleFetchDirectoryBackup,
it is safe to delete it after ceph-ansible has run. If we need
to apply a system update, the fetch directory will be restored
each time before ceph-ansible is run and it will then subsequently
be backed up and deleted.

During the restore, problems can occur if the update is being run
by a different user or if the restore tries to overwrite metadata.
Rather than try to handle this by setting permissions or passing
metadata overwrite flags, it is simpler to just ensure that each
restore is a restore to a new copy of the fetch directory.

Also, remove the task which removes nested sections of fetch
directory (if any) and remove registrations of nested sections.
This task's goal was to clean the fetch directory which would
continue to exist on the filesystem. Since that directory is
now being fully removed there is nothing to clean.

Change-Id: I7a01146a135254ccb948b8e793859f8fa3bdc690
Closes-Bug: #1823229