standalone/undercloud - post: use EndpointMap to fetch Keystone URL
Using EndpointMap to ensure we get the hostname/fqdn if possible
otherwise it fallbacks to the IP for Keystone public endpoint.

This is useful when the operator uses a certificate based on
hostname/fqdn and not an IP address.

Closes-Bug #1763776

(cherry picked from commit 016279b71e74700c18f2270649cb7567ae94505f)
Depends-On: I1c222fce178e164432acbfaeda8695c3cf7a6e98 
Change-Id: Id8e1c6408ee6a322c61de90a52ab1eacaf0dba88