Run nova-manage as root to prevent wrong nova-manage.log permissions
To verify if the default cell is created we run a nova-manage command.
Right now this is triggered as root and not as nova like the otheres.
In case it is the first time on the node it creates the nova-manage.log
with root user and follow up call as nova user will fail.

Related-bug: 1820590

Change-Id: I56a961615c0afa748a42c25d4614637275b7c33b