ensure /var/run/redis is present upon reboot
/var/run is mounted as a tmpfs, which means /var/run/redis
disappears on reboot. Since podman doesn't try to recreate a
directory if it doesn't exist, containerized redis cannot
be restarted because one of its bind mount is missing.

Force the creation of /var/run/redis after every reboot, with
default permissions. On container restart, Kolla will chown and
chmod the directory to the proper value.

Change-Id: I36c81f58c3b3e9366212a93f73b6212cca4ea569
Closes-Bug: #1826554
Depends-On: Iaa8a99eb9ced21fb6c7c87c5b56dec55383af9a9