Switch off nova metadata api on the undercloud
All known consumers of boot data (os-collect-config, etc) have a
preference for using config-drive as the data source.

The last known consumer was novajoin, but that switched to preferring
config-drive early in the Stein development cycle[1] so it should now
be safe to switch off the nova metadata API service.

[1] https://review.opendev.org/#/c/607492/
Blueprint: nova-less-deploy

Change-Id: If35aec24f446769fca7897c2126fb6657454f073