Change setype for non container /var/log/nova-placement
I870f6f5286ac9222d6d8fbdcb48fe6b6559993ba changes the setype to
var_log_t for the other services. Since nova-placement was
removed in train the backport does not change it.

This is a stein and lower fix to change it also for nova-placment.

/var/log/<service> is not used by containers, therefore setype
svirt_sandbox_file_t is not required. This changes it to the
default var_log_t.

Change-Id: I3d7b43db67434c8a7331c712f04b61cff14b993e
(cherry picked from commit 4e36f4918399d4910548b77f1721903239e4baca)