Clean up leftover mount point after docker stop.
Some mountpoints are left on the host after docker shut down.  This
seems to be an issue with docker, but couldn't trace it back to a
specific docker bug.

In the meantime to unblock the upgrade CI we make sure that every
mount point under /var/lib/docker are umounted before deleting that
directory.

Note that we need to keep the order right, so that we do a depth first
list so that umount can do leaves before root.

Closes-Bug: #1826375
Co-Authored-By: Sofer Athlan-Guyot <sathlang@redhat.com>
(cherry picked from commit a8ec69941657975fe6a2b8616c3c729013a30dff)

Change-Id: I03a065556caca4385bb8b28be0dfbe21addbf003