Add in-flight health validations to Ceph deployment
Display a warning if the Ceph cluster health is not
OK using the validations framework. If the ceph health
is in a state of error, then fail the deployment and
display a message saying the same to save deployment
time.

Add new CephOsdPercentageMin parameter which fails the
deployment early if a certain percentage of OSDs are not
running and in the Ceph cluster. This parameter defaults
to 66%; if less than 2/3rd of the OSDs are not in after
Ceph is deployed then fail the deployment early so the
user may debug. Setting CephOsdPercentageMin to 0
disables this check.

Change-Id: I62251007b311f4cc27afc7759787d9c82bf49957
Implements: blueprint inflight-validations
Depends-On: https://review.opendev.org/#/c/657173