Set configure_delegated_roles a parameter
The configure_delegated_roles configure if the heat_stack_onwer role
would be created or not by keystone. Right now this is set to false,
without any way to override to true. This patch change this option to be
a parameter and it also change scenario001 to true in order to run heat
tempest tests.

Change-Id: I916cc4842ccef587a25b06cb422436953974e790