Gracefully handle empty config json data
If the json config data is set to Null, the current code throws an error
because we attempt to use .items() on the parsed result. We should just
treat and empty file as an empty dict so the code properly continues.

Change-Id: I02bb2ce5359aeec1e600b0c0f04fb72c7a335407
Closes-Bug: #1828295
(cherry picked from commit 055d15f9288f7c49670c47e847ead564d8b2d5de)