Remove OpenStack packages right before upgrading the operating system
This could make the upgrade faster, but mainly it's less likely that
something will go wrong when the node goes down for the offline
upgrade. We want to minimize the chances of that happening, because
such error is extremely unpleasant to deal with (restore from backup
might be the only solution at that point).

Change-Id: Iddf30278f0cb4eb93fcbd8dc829e247bfb534562