Remove deployed-server bootstrap resource
Removes the deployed-server bootstrap resource from the deployed-server
template as the equivalent functionality enabled by the scripts has been
moved to the tripleo-bootstrap ansible role provided by tripleo-common.

The environments and templates for the bootstrap resource are deprecated
in train and will be removed in a future release.

Change-Id: If5adc76190e986bae1346a3a7e8dda8a2bc92f27
Depends-On: Idc13a88481ae618f0321009f49acf71e68258b95
implements: blueprint reduce-deployment-resources