Make NovaResumeGuestsStateOnHostBoot podman compatible
With podman therefore are changes required to the libvirt-guests
unit file and how we dealt with dependencies of the containers.
With podman systemd is used to manage the containers which we now
can use for container dependencies and use libvirt-guests from
inside a container.

Change-Id: I2a1f370732effd3303942dd5ec46892ec4a85df9
Closes-Bug: #1830067
(cherry picked from commit 6b0d025fd6f5b44428555d12071c700d794a687e)