Use make_url to wrap IPv6 addresses in brackets.
Replace the dependency of network data values for wrapping
'ip_address_uri' in brackets. Instead of using jinja2
variables network.ipv6 or ipv6_override to decide if the
IP need to be wrapped in brackets or not the make_url
function in heat in combination with a str_replace is
used.

Related-Bug: #1830406
Change-Id: I2d8c405c1df30ac11cc2286398fe4749694da10e