Add CephAnsibleEnvironmentVariables to nodes-uuid call
CephAnsibleEnvironmentVariables are also useful when running
the nodes-uuid playbook. Users may know ceph-ansible playbook
is run but may not know the nodes-uuid playbook is run too.
If additional Ansible environment variables are useful for
running ceph-ansible it is likely they will be needed for
the nodes-uuid playbook. The altnernative is to create another
parameter like NodesUuidAnsibleEnvironmentVariables.

Change-Id: I10ddb4f79f5c8b69b09622b96e96325ba19d62e0