Convert tuned puppet to ansible
The tuned heat template was using puppet to configure the service.
This change removes `tuned-baremetal-puppet.yaml` and replaces it
with `tuned-baremetal-ansible.yaml` in an effort to decrease our
reliance on puppet.

The old puppet file name was removed from the resource registry
template and replaced with the new file name.

Additional changes were made to the tuned boot params template
and tasks. These changes modify the tasks for readability and
ensure we're not needlessly running shell commands. Because the
parameter `TunedProfileName` is now used in the new ansible
based templatei, changes have been made to ensure we're enforcing
a consistent definition.

> Once this change has been merged a follow-up patch will be made
  to puppet-tripleo to eliminate the tuned.pp manifest.

Change-Id: I06d07f6f0949095cb716dd706e05f2e567c0d3d8
Signed-off-by: Kevin Carter <kecarter@redhat.com>