Handle openvswitch meta-package on undercloud upgrade.
As done for the overcloud in [0], we need to include the
same special package handling in the tasks in charge of
upgrading the undercloud.

This patch adds special handling for the rhosp-openvswitch
layered product package to ensure that openvswitch networking
remains available during the upgrade.

[0] - https://opendev.org/openstack/tripleo-heat-templates/commit/1773afb0689c8fded9bce2dff4c1f8d9406c8294

(cherry picked from commit 0d9d92ac58aeae25b8c120ab8ebc2bad7233d9bc)
Change-Id: Ib115e70ce0ce53a8638ccbe6c9fe6ed6a421e09b
Closes-Bug: #1830627