Change package installs for Stein and OS upgrade
After upgrading to a new OS we might not have all the necessary
packages, so install them. The list was taken from deployed-server
preparation script. Also change the UpgradeInitCommonCommand for
Stein. e.g. we no longer require Ansible and Ansible-Pacemaker on the
overcloud, as Ansible is now being executed from undercloud only.

Change-Id: I7381423b9834d6ebe6e29a07fb3996ca79398187
Closes-Bug: #1830882