krb-service-principals support service_net_map_replace
Handle service_net_map_replace in the jinja2 logic so that
service_net_map_replace works for both default networks and
custom networks.

Enables a user either to change 'name_lower' of a network
and overriding the ServiceNetMap accordingly, as well as
user changeing 'name_lower' and use 'service_net_map_replace'
so that the default ServiceNetMap can be used.

Related-Bug: #1830852
Change-Id: Iae4341e9e7c888da4dd8d0dedd5ad28b7e0e6c40