Remove unnecessary openldap-clients package from overcloud controllers
We currently install the openldap-clients package on overcloud
controller nodes.  It was assumed that this package was needed for
IPA client enrollment, but the ipa-client package only uses the
openldap libraries that will already be pulled in as an indirect
dependency of ipa-client.  The openldap-clients package contains
LDAP CLI utilities, which don't appear to be actually used.

Change-Id: I14c1d8204bd84ead0d6995b5aefd10d2bbc4227d