Enabling Neutron Routed Provider Networks (segments)
The 'segments' plugin will be added to the NeutronServicePlugins.

Neutron routed networks was enabled by default[1], with the switch
to OVN it was removed[2] from the list of default service plug-ins.

It makes sense to enable the plug-in by default so that it's easy
to scale out existing provider networks without being limited by
the restrictions[3] related to scaling existing provider networks
that was created prior to enabling the service_plugin.

[1] https://review.opendev.org/562932
[2] https://review.opendev.org/593056
[3] https://opendev.org/openstack/neutron/src/branch/master/releasenotes/notes/allow-update-subnet-segment-id-association-1fb02ace27e85bb8.yaml

See for details:
https://docs.openstack.org/neutron/latest/admin/config-routed-networks.html

Co-Authored-By: Harald Jensås <hjensas@redhat.com>
Closes-Bug: #1831969
Change-Id: Ie67134719abc3b89dff19130e1557517bdc18714
Related-blueprint: tripleo-routed-networks-deployment
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>