Add new role parameter NovaLibvirtNumPciePorts
Add role parameter NovaLibvirtNumPciePorts which sets `libvirt/num_pcie_ports`
to specify the number of PCIe ports an instance will get.
Libvirt allows a custom number of PCIe ports (pcie-root-port controllers) a
target instance will get. Some will be used by default, rest will be available
for hotplug use. When using the 'q35' machine type, by default, it allows only
a *single* PCIe device to be hotplugged.  And Nova currently sets
'num_pcie_ports' to "0" (which means, it defaults to libvirt's "1"), which is
not sufficient for hotplug use.

Default for NovaLibvirtNumPciePorts is 16.

Change-Id: Ida27b52a091640545aecc982fc1a509fb5107db8
Closes-Bug: #1831701
Depends-On: I16732c9d6013112381cfad999540dd41ec3d7ba3