Add networks to Standalone role
The standalone role can be used with and without an undercloud. If you
deploy it with an undercloud and use network-isolation, then the
deployment will fail because the networks aren't defined in the role.
Since it doesn't hurt to have them on in the non-undercloud version,
lets add them back in.  They were previously removed as part of
https://review.opendev.org/#/c/628265/

Change-Id: I366631cb6ecd39595c00204a02408beb431af70e
Closes-Bug: #1832043
(cherry picked from commit da1827f7d348fd90e00f8381347febe01b6f7239)