Fix network rename + service_net_map_replace
If name_lower of a network is changed, and service_net_map_replace
is used. The subnet defined for the network in roles_data would
have to be updated to match the new name_lower. This was not the
case prior to stein, because roles_data did'nt include the subnet.

A downstream regression test is failing because of this. The
workaround is to update roles_data with a value for the subnet
that match name_lower.

This patch should fix the need to update roles_data in the case
that the net_map_replace matches the subnet in roles_data and no
additional subnet with this defined for this network in
network_data.

Closes-Bug: #1832759
Change-Id: I10f99fad0f0016bf81edd53e1c5579f9d5c0ef88