Enable Request Filter for Image Types
This change enables `scheduler/query_placement_for_image_type_support`
by default for all deployments. Setting it causes the scheduler to ask
placement only for compute hosts that support the disk_format of the
image used in the request which is beneficial for example, the libvirt
driver, when using ceph as an ephemeral backend, does not support ``qcow2``
images (without an expensive conversion step).

Change-Id: I6d12a66616cc2cc65e62755e8a54084d65aeae5e
Closes-Bug: #1832738