Add new parameter NovaLiveMigrationWaitForVIFPlug
Add parameter NovaLiveMigrationWaitForVIFPlug which allows to set
`compute/live_migration_wait_for_vif_plug` which in turn allows
whether to wait for ``network-vif-plugged`` events before
starting guest transfer.

The default value for the parameter is set to true and it is
set to false when `ovn` is in `NeutronMechanismDrivers` list.

Depends-On: I0048d1f57eeb3418b52f225f87810ea1a7250a0f
Change-Id: I1ce38ea3ef6ba40625adc8da90903aac397cf60e