Add container engine authentication support
We need to be able to run a docker or podman login during the
deployment if the overcloud needs to pull images from an auth
required container registry when deploying with docker. Add
ContainerImageRegistryLogin as a flag to use
ContainerImageRegistryCredentials to perform docker or podman logins
when deploying.

Closes-Bug: #1833584
Change-Id: I98a527f363056767fea45ab4828ae61c01de20ca
(cherry picked from commit ea37ee6638b08375974ae3ceddd6dc136c67b983)