Configure the Multipathd service
This is part 2 of a series of patches to properly deploy multipathd.

This patch configures /etc/multipath.conf, which is required by the
multipath daemon. The config file contents are managed using the
'mpathconf' tool instead of a puppet module, but a tiny puppet wrapper
is employed to take advantage of the tripleo tooling (e.g. restarting
the multipathd container whenever the config file changes).

Other changes include:
- Ensuring the multipath daemon isn't running on the host
- Preparing the /etc/multipath directory that must be shared by all
  containers that use the service
- Adding a volume mount to services that need access to /etc/multipath

Closes-Bug: #1834042
Change-Id: Ib3ce0a753f7220c3ee2556425e9798b0c4a906f3