Configure Max Delay for purge job about shadow tables in nova
Recently we have a cron job to purge records in nova shadow tables.
This cron job runs on all controller nodes, but as all controllers
have configuration to run the job at the really same time, which
can result in db lock timeout or db spike.

This patch makes sure that we set a default value to Max Delay
parameter for the cron job, to avoid the problem caused by the
collision.

Change-Id: Id372ce98d8ba761ef313bd90a406737250503cab