Allow for configuration of the Manila CephFS backend with a remote Ceph
When deploying Manila with CephFS backend it might be preferable
to keep the Ganesha service deployment under control of TripleO but
still point it to an external Ceph cluster.

From version 4.0.0rc9 [1] ceph-ansible is capable of deploying Ganesha and
configure it to use an external Ceph cluster. We can use this feature
TripleO enabling the CephNfs and CephClient services on those roles
where Ganesha needs to go.

It will be possible to deploy Manila with a CephFS backend together
with Ganesha (pacemaker managed) and point them to an external Ceph
cluster combining the two environments:
- manila-cephfsganesha-config.yaml
- ceph-ansible/ceph-ansible-external.yaml

1. https://github.com/ceph/ceph-ansible/commit/6a6785b719d126cf54ebad8b2a22c97d90afd05e

Change-Id: Iae674127cf948fbc007ee2be96041221f7584ce7
Implements: blueprint manila-external-ganesha