CI should auto-generate server_certs_key_passphrase
Bug 1833942 showed that in a case that the generated value
server_certs_key_passphrase is invalid, Octavia will fail to operate.

In CI, we currently provide a pre-defined passphrase that might cover
for potential breakages in the future. This patch removes the
pre-defined passphrase so it will get generated on each run.

Note that, TripleO will now[1] either auto-generate a valid passphrase
or validate a pre-defined one.

Related-Bug: #1833942

[1] https://review.opendev.org/#/q/topic:OctaviaServerCertsKeyPassphrase-32chars

Depends-On: I5c2629d9e7700fe1dd6f915bc257b1f058e40617
Depends-On: Ibcdbe2605a7cabe3a5ef8245b4460c8f70220989
Depends-On: I886f2b8ac7092d9b3da38852e92a615d5666eea7

Change-Id: Ie596b04614c2ca9d961694f4012c1553a092aa3e