Disable Docker in default Standalone environments
Podman is the default in standalone generated environments
(e.g. environments/standalone/standalone-tripleo.yaml), however since we
haven't made it the default in overcloud-resource-registry-puppet.j2.yaml
until we get CentOS8, docker was still being deployed because the
roles/Standalone.yaml used to contain the Docker service.

This patch aims to make sure we disable Docker.
Note: for scenario004 & 012, we need to enable Docker as Pacemaker is
enabled and the job runs on CentOS7.

Closes-Bug: #1835411
Change-Id: Ib34ba24c84f34a1533a90189d5154825c6dfa868