Don't use the z flag in case NovaNfsEnabled is true
Deployment with enabled NFS share for nova ephemeral storage fails. Podman
fails to relable with mounted nfs in /var/lib/nova/instances and container
fail to start with "operation not supported".
This change only sets the z flag for the /var/lib/nova in case nfs is not
enabled for the compute.

Change-Id: I732c0c3e0b3eb1b52f0df58568ec3a42f3d6d1a1
Closes-Bug: #1835503