Handle edge cases in staged upgrade hiera data
We make sure the hiera is set in an either-or branches so that we
prevent from executing both at the same time or none at all.

Edge case 1 which is not that much of an edge case --
single-controller environment with Pacemaker. Previously we ran both
branches resulting in duplicate entries.

Edge case 2 -- having the service in roles data but having no node
assigned to that role. In that case previously we wouldn't set the
variables and the upgrade would fail.

Change-Id: I1681a5e5ac7842500fc5e9dd0900ee6121773537
Partial-Bug: #1831690
(cherry picked from commit 2b427970a36a63cc2b9afbd61f10fe033287aca5)