Use template for NetworkConfig script
Use the first_found Ansible lookup plugin to look for the NetworkConfig
script in both the role directory and server directory. This allows for
using a role generic NetworkConfig script in the role dir, while still
being backwards compatible.

The NetworkConfig script is also rendered from a template on the remote
node instead of just copied. Treating it as a template allows it to be
made generic per role instead of node specific.

Depends-On: I6bbca8eb256f387c4f74378baa5253890cfe1abf
Change-Id: Ibbb5dd2ccbdaec083ec636ac44dfdc75b65f89a7