Specify a default for container_registry_logins
When we run this, we'll run the controller host_prep_tasks but the
compute nodes are included. Ansible will skip the set_fact on the
computes but still try to evaluate the lookup. We need to use |default
to handle the unset case.

Change-Id: I1ff41befbd66325d90c3ea18640a298549d1b659
Closes-Bug: #1835657
(cherry picked from commit a6ff00885ce0540015b0ed1fc99c431a87348077)